moving status updater to utils
